requirement.txt
Flask-Migrate
Flask-SocketIO
Flask-Caching
Flask-Login
marshmallow
PyJWT
python-dotenv
redis
gunicorn 